Adaptor metrics dev.
git-svn-id: https://openadaptor3.openadaptor.org/svn/openadaptor3/trunk@1771 c0ec5643-5123-0410-b01b-e101a1959115

+++ src/org/openadaptor/core/recordable/IRecordableComponent.java
@@ -39,11 +39,11 @@ public interface IRecordableComponent {
   
 	long getMessageCount();
 	
-	float getMaxProcessTime();
+	long getMaxProcessTime();
 	
-	float getMinProcessTime();
+	long getMinProcessTime();
 	
-	float getAvgProcessTime();
+	long getAvgProcessTime();
 
 	String getMessageType();
 }
